+++ home/moz4r/AIML/FRENCH/update.aiml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<aiml>
+<!--  -->
+<!--  -->
+<category><pattern>PROGRAM VERSION</pattern>
+<template>Program AB 0.0.6.26 beta -- AI Foundation Reference AIML 2.1 implementation</template>
+</category>
+<category><pattern>BRAIN BUILD</pattern>
+<template>Thu Jun 30 02:11:54 CEST 2016</template>
+</category>
+</aiml>
