+++ home/moz4r/README
@@ -1,0 +1,23 @@
+Rachel BOT is french/english AIML based bot
+Sorry english version is just a POC but fully worky, I work on french AIML in priority
+ToDo : Inmoov gesture
+
+# ##############################################################################
+# 							*** SETUP ***
+# ##############################################################################
+# STABLE FILES : https://github.com/MyRobotLab/aiml/tree/master/bots/ [ RACHEL AIML + PYTHON ]
+# UPDATED DEV FILES : https://github.com/MyRobotLab/pyrobotlab/tree/master/home/moz4r [ RACHEL AIML + PYTHON ]
+# -----------------------------------
+# Rachel bot Version 1.2 By Moz4r
+# 
+# Wikidatafetcher By Beetlejuice
+#
+# french AIML dictionay with the help of MAT
+# -----------------------------------
+# !!! INSTALL : !!!
+# !!! Copy all aiml files to : develop\ProgramAB\bots\rachel\aiml !!!
+# !!! All Rachel*.py to the root of MRL
+# !!! https://github.com/MyRobotLab/aiml/tree/master/bots/BOTS-FRENCH/Rachel/TXT to the root of MRL
+# !!! https://github.com/MyRobotLab/aiml/tree/master/bots/BOTS-ENGLISH/Rachel/TXT
+# !!! 
+# You can use mrl.bat to launch, it's usefull to kill java and autostart the script
