--- home/moz4r/learn.aiml
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<aiml>
-<!--  -->
-<!--  -->
-<category><pattern># APPRENTISSAGE #</pattern>
-<template><random>
-<li>J'aime apprendre de nouvelle chose, je t'écoute</li>
-<li>Je suis prêt à apprendre</li>
-</random>
-<think><set name="topic">LEARN1</set></think></template>
-</category>
-<topic name="LEARN1">
-<category><pattern>UN * EST UN *</pattern>
-<template><random>
-<li>Ok, un <star/> est un <star index="2"/>.</li>
-<li>Maintenant je sais que un <star/> est un <star index="2"/>.</li>
-</random>
-<oob>
-			<mrl>
-				<service>python</service>
-				<method>exec</method>
-				<param>SaveMemory()</param>
-			</mrl>
-      </oob>
-<learn>
-<category><pattern>C'EST QUOI UN <eval><star/></eval></pattern>
-<template>
-UN <eval><star/></eval> est UN <eval><star index="2"/></eval>.</template>
-</category>
-</learn>
-<learn>
-<category><pattern>EST-CE QUE UN <eval><star/></eval> EST UN <eval><star index="2"/></eval></pattern>
-<template>
-OUI un <eval><star/></eval> est un <eval><star index="2"/></eval>.</template>
-</category>
-</learn></template>
-</category></topic>
-
-
-<topic name="LEARN1">
-<category><pattern>EST-CE QUE UN * EST UN *</pattern>
-<template>Non ! Un <star/> n'est pas un <star index="2"/>!</template>
-</category></topic>
-
-</aiml>
